// src/app/models/projectProviders.ts
export interface ProjectProviders {
    $uid?: string;
    $key?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    phone?: string;
    organization?: string;
    position?: string;
    password?: string;
    experience?: string[];
    expertise?: string;
    //post
    lookingFor?: string;
    studentMajor?: string;
    studentYear?: string;
    studentSemester?: string;
    studentSkills?: string;
    studentExperience?: string;
    projectType?: string;
    projectName?: string;
    projectDescription?: string;
    projectDuration?: string;
    projectLocation?: string;
    paidOrUnpaid?: string;
    comments?: string;
    
}